4-rag-with-vectordb/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.18.0
fsspec==2025.5.1
google-generativeai==0.8.5
grpcio==1.67.1
huggingface-hub==0.32.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf>=3.20.2,<6.0.0
pymilvus==2.5.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
